# perf(ReportPDF): RHICOMPL-3214 make querying failed rules more efficient