The same data is preloaded on a lower level, therefore it's more efficient to query it there. Should reduce the number of allocations significantly.

```
Completed 200 OK in 1483ms (Views: 0.7ms | ActiveRecord: 1264.1ms | Allocations: 151263)
Completed 200 OK in 1421ms (Views: 0.5ms | ActiveRecord: 1255.3ms | Allocations: 71637)
```

Tested by comparing two PDFs from the old and the new codebase and the results were identical.

## Secure Coding Practices Checklist GitHub Link
- https://github.com/RedHatInsights/secure-coding-checklist

## Secure Coding Checklist
- [ ] Input Validation
- [ ] Output Encoding
- [ ] Authentication and Password Management
- [ ] Session Management
- [ ] Access Control
- [ ] Cryptographic Practices
- [ ] Error Handling and Logging
- [ ] Data Protection
- [ ] Communication Security
- [ ] System Configuration
- [ ] Database Security
- [ ] File Management
- [ ] Memory Management
- [x] General Coding Practices